backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
celery[redis]==5.3.6
redis==5.0.1
pillow==10.2.0
pytesseract==0.3.10
opencv-python-headless==4.9.0.80
numpy==1.26.4
python-dotenv==1.0.1
pydantic[email]==2.6.1
email-validator==2.1.0
pydantic-settings==2.1.0
httpx==0.26.0
pytest==7.4.4
pytest-asyncio==0.23.4
